Implementa método para pesquisar resumos de estoque na interface EstoqueRepositorioAplicacao e no serviço EstoqueServicoAplicacao. Atualiza o controlador EstoqueControlador para expor a nova funcionalidade via API, permitindo a listagem de resumos de estoque. Altera o tipo de peso na classe ProdutoJpa de Double para BigDecimal para melhor precisão.

+++ infraestrutura/src/main/java/dev/gestock/sge/infraestrutura/persistencia/jpa/ProdutoJpa.java
@@ -1,5 +1,6 @@
 package dev.gestock.sge.infraestrutura.persistencia.jpa;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -38,7 +39,7 @@ class ProdutoJpa {
 	String unidadePeso;
 
 	@Column(nullable = false, precision = 10, scale = 3)
-	Double peso;
+	BigDecimal peso;
 
 	@Column(nullable = false)
 	Boolean perecivel;
